demo/app/requirements/common.txt
matplotlib==3.5.0
opencv-python==4.5.4.58
easyocr==1.4.1
numpy==1.21.4
pandas==1.3.4
keras-ocr==0.8.8
tensorflow==2.7.0
vietocr==0.3.7
fastapi==0.70.0
Pillow==8.2.0